+++ enterprise/cmd/frontend/internal/batches/resolvers/batch_change_connection.go
