+++ enterprise/cmd/frontend/internal/batches/resolvers/bulk_operation.go
