+++ enterprise/cmd/frontend/internal/batches/resolvers/errors.go
